+++ back-end/src/database/models/product.js
@@ -1,10 +1,16 @@
 const Product = (sequelize, DataTypes) => {
   const newProduct = sequelize.define('product', {
     name: DataTypes.STRING,
-    price: DataTypes.DOUBLE,
+    price: DataTypes.DECIMAL(5,2),
     url_image: DataTypes.STRING
   },
   { timestamps: false });
+
+  // tirar daqui
+  // newProduct.associate = (models) => {
+  //   newProduct.hasMany(models.salesProducts,
+  //     { foreignKey: 'product_id', as: 'products' });
+  // }
 
   return newProduct;
 };
